+++ src/components/KlabMainControl.vue
@@ -30,7 +30,7 @@
     <q-card
       id="mc-q-card"
       class="no-box-shadow absolute lot-of-flow"
-      :class="[hasContext ? '' : 'bg-transparent', hasContext ? 'with-context' : 'without-context']"
+      :class="[hasContext ? 'with-context' : 'bg-transparent without-context', largeMode ? 'mc-large-mode' : '']"
       :style="{ top: `${defaultTop}px`, left: `${centeredLeft}px` }"
       :flat="true"
       v-draggable="dragMCConfig"
@@ -41,8 +41,9 @@
         id="mc-q-card-title"
         class="q-pa-xs"
         ref="mc-draggable"
+        :class="[ fuzzyMode ? 'klab-fuzzy' : '', searchIsFocused ? 'klab-search-focused' : '']"
         :style="{
-          'background-color': getBGColor(hasContext ? '1.0' : searchIsFocused ? '.6' : '.2'),
+          'background-color': getBGColor(hasContext ? '1.0' : searchIsFocused ? '.8' : '.2'),
         }"
       >
         <klab-search-bar ref="klab-search-bar"></klab-search-bar>
@@ -113,11 +114,10 @@
 </template>
 
 <script>
-// import Vue from 'vue';
 import { mapGetters, mapActions } from 'vuex';
 // import { Draggable } from 'draggable-vue-directive';
 import { Draggable } from 'shared/VueDraggableTouchDirective';
-import { VIEWERS, CUSTOM_EVENTS, LEFTMENU_CONSTANTS } from 'shared/Constants';
+import { VIEWERS, CUSTOM_EVENTS, LEFTMENU_CONSTANTS, CONSTANTS } from 'shared/Constants';
 import { dom, debounce } from 'quasar';
 import MainActionsButtons from 'components/MainActionsButtons';
 import StopActionsButtons from 'components/StopActionsButtons';
@@ -178,18 +178,21 @@ export default {
     ...mapGetters('data', [
       'hasContext',
     ]),
+    ...mapGetters('stomp', [
+      'hasTasks',
+    ]),
     ...mapGetters('view', [
       'spinnerColor',
       'searchIsFocused',
       'isDrawMode',
-    ]),
-    ...mapGetters('stomp', [
-      'hasTasks',
+      'fuzzyMode',
+      'largeMode',
     ]),
   },
   methods: {
     ...mapActions('view', [
       'setMainViewer',
+      'setLargeMode',
     ]),
     callStartType(event) {
       if (!this.searchIsFocused) {
@@ -274,6 +277,7 @@ export default {
   },
   watch: {
     hasContext() {
+      this.setLargeMode(false);
       this.$nextTick(() => {
         this.changeDraggablePosition({
           top: this.defaultTop,
@@ -281,6 +285,18 @@ export default {
         });
       });
       // this.draggableElement.classList.remove('vuela');
+    },
+    largeMode() {
+      if (this.hasContext) {
+        return;
+      }
+      this.$nextTick(() => {
+        const offset = CONSTANTS.SEARCHBAR_INCREMENT * (this.largeMode ? 1 : -1) / 2;
+        this.changeDraggablePosition({
+          top: parseFloat(this.draggableElement.style.top),
+          left: parseFloat(this.draggableElement.style.left) - offset,
+        });
+      });
     },
   },
   created() {
@@ -311,24 +327,22 @@ export default {
 
   #mc-container
     #mc-q-card-title
-    .q-card
-      width $main-control-width
-      &.with-context
-        #mc-q-card-title
-          width $main-control-width - 30px
-
-    #mc-q-card-title
       border-radius 30px
       cursor move
       transition background-color 0.8s
     .q-card
       overflow auto
+      width $main-control-width
+      transition width .5s
       &.with-context
+        width $main-control-width - 30px
         background-color rgba(35, 35, 35 ,.8)
         border-radius 5px
         #mc-q-card-title
           overflow hidden
           margin 15px
+      &.mc-large-mode
+        width $main-control-width + $main-control-inc-width
 
     .q-card-title
       position relative