+++ src/pages/index.vue
@@ -131,6 +131,7 @@ export default {
       'mainViewerName',
       'mainViewer',
       'isInModalMode',
+      'spinnerErrorMessage',
     ]),
     logVisible() {
       return this.$logVisibility === WEB_CONSTANTS.PARAMS_LOG_VISIBLE;
@@ -236,7 +237,7 @@ export default {
         event.preventDefault();
         return;
       }
-      if (event.keyCode === 38 || event.keyCode === 40 || event.keyCode === 32 || this.isAcceptedKey(event.key)) {
+      if (event.keyCode === 9 || event.keyCode === 38 || event.keyCode === 40 || event.keyCode === 32 || this.isAcceptedKey(event.key)) {
         if (!this.searchIsActive) {
           this.searchStart(event.key);
           event.preventDefault();