+++ CHANGELOG.md
@@ -7,18 +7,20 @@ is not available.
 
 **The version of k.Explorer software is self-referred and is not linked to the k.LAB version, so is possible that some features are not available in the stable version of the engine.**
 
-## 0.7.5 - 2019-06-27
+## 0.7.5 - 2019-07-01
 ### Added
 - with double click on map, map info popup is locked and show info of the top layer. To put on top a layer, 
 you can double click in the tree on it.
 - added dataflow element information on click
 - the search bar has new feature for more complex queries
+- possibility of free search
 ### Improvements
 - with a click on map, info about the point is showed even if observation info panel is open
 - add a separator in log visualizer for each context reset
 - main action buttons redesign
 - improvement on dataflow viewer
 - notifications of observation between explorer and engine was completely redefine
+- spacebar to select element
 ### Solve
 - unity of scale bug solved
 - lock scale functionality now is coherent